0.5.2 Patched the database, fixing the API

+++ app/src/main/java/com/useless/boosterapp4/data/MoviesDatabase/MovieMapper.kt
@@ -12,7 +12,7 @@ class MovieMapper {
                 movieResponse.Lang,
                 movieResponse.movie_title,
                 movieResponse.date,
-                movieResponse.voteAvg,
+                movieResponse.voteAvg.toInt(),
                 movieResponse.voteCnt,
                 movieResponse.overview,
                 movieResponse.id
